moved tests to groovy
