+++ src/test/groovy/com/github/sth/groovy/vertx/mongo/streams/GridFSInputStreamTest.groovy
@@ -1,16 +1,15 @@
-package com.github.sth.vertx.mongo.streams;
+package com.github.sth.groovy.vertx.mongo.streams
 
-import com.github.sth.vertx.mongo.streams.util.ByteUtil;
-import com.github.sth.vertx.mongo.streams.util.DrainHandler;
-import com.github.sth.vertx.mongo.streams.util.ResultCallback;
+import com.github.sth.groovy.vertx.mongo.streams.util.ByteUtil;
+import com.github.sth.groovy.vertx.mongo.streams.util.DrainHandler;
+import com.github.sth.groovy.vertx.mongo.streams.util.ResultCallback;
 
 import org.junit.Assert;
 import org.junit.Test;
 
-import java.nio.ByteBuffer;
-import java.util.Arrays;
+import java.nio.ByteBuffer
 
-import io.vertx.core.buffer.Buffer;
+import io.vertx.groovy.core.buffer.Buffer;
 
 public class GridFSInputStreamTest {
 
@@ -22,8 +21,8 @@ public class GridFSInputStreamTest {
   public void happyPathReadAfterWrite() {
     GridFSInputStream inputStream = GridFSInputStream.create();
 
-    Buffer buffer1 = Buffer.buffer(ByteUtil.randomBytes(2048));
-    Buffer buffer2 = Buffer.buffer(ByteUtil.randomBytes(2048));
+    Buffer buffer1 = new Buffer(io.vertx.core.buffer.Buffer.buffer(ByteUtil.randomBytes(2048)));
+    Buffer buffer2 = new Buffer(io.vertx.core.buffer.Buffer.buffer(ByteUtil.randomBytes(2048)));
 
     // write data to the stream
     inputStream.write(buffer1);
@@ -40,8 +39,8 @@ public class GridFSInputStreamTest {
     Assert.assertTrue(resultCallback.succeeded());
     Assert.assertEquals(4096, resultCallback.getResult(), 0);
 
-    Assert.assertTrue(Arrays.equals(buffer1.getBytes(), Arrays.copyOfRange(byteBuffer.array(), 0, 2048)));
-    Assert.assertTrue(Arrays.equals(buffer2.getBytes(), Arrays.copyOfRange(byteBuffer.array(), 2048, 4096)));
+    Assert.assertTrue(Arrays.equals(buffer1.delegate.getBytes(), Arrays.copyOfRange(byteBuffer.array(), 0, 2048)));
+    Assert.assertTrue(Arrays.equals(buffer2.delegate.getBytes(), Arrays.copyOfRange(byteBuffer.array(), 2048, 4096)));
 
     // on the next invocation the mongo driver should be signaled that no more data is available
     resultCallback = new ResultCallback<>();
@@ -59,8 +58,8 @@ public class GridFSInputStreamTest {
   public void happyPathReadBetweenWrites() {
     GridFSInputStream inputStream = GridFSInputStream.create();
 
-    Buffer buffer1 = Buffer.buffer(ByteUtil.randomBytes(2048));
-    Buffer buffer2 = Buffer.buffer(ByteUtil.randomBytes(2048));
+    Buffer buffer1 = new Buffer(io.vertx.core.buffer.Buffer.buffer(ByteUtil.randomBytes(2048)));
+    Buffer buffer2 = new Buffer(io.vertx.core.buffer.Buffer.buffer(ByteUtil.randomBytes(2048)));
 
     // write data to the stream
     inputStream.write(buffer1);
@@ -73,7 +72,7 @@ public class GridFSInputStreamTest {
 
     Assert.assertTrue(resultCallback.succeeded());
     Assert.assertEquals(2048, resultCallback.getResult(), 0);
-    Assert.assertTrue(Arrays.equals(buffer1.getBytes(), Arrays.copyOf(byteBuffer.array(), 2048)));
+    Assert.assertTrue(Arrays.equals(buffer1.delegate.getBytes(), Arrays.copyOf(byteBuffer.array(), 2048)));
 
     // write data to the stream
     inputStream.write(buffer2);
@@ -83,7 +82,7 @@ public class GridFSInputStreamTest {
 
     Assert.assertTrue(resultCallback.succeeded());
     Assert.assertEquals(2048, resultCallback.getResult(), 0);
-    Assert.assertTrue(Arrays.equals(buffer1.getBytes(), Arrays.copyOf(byteBuffer.array(), 2048)));
+    Assert.assertTrue(Arrays.equals(buffer1.delegate.getBytes(), Arrays.copyOf(byteBuffer.array(), 2048)));
 
     // on the next invocation the mongo driver should be signaled that no more data is available
     resultCallback = new ResultCallback<>();
@@ -107,19 +106,19 @@ public class GridFSInputStreamTest {
     // empty queue should not be full
     Assert.assertFalse(inputStream.writeQueueFull());
 
-    Buffer buffer = Buffer.buffer(ByteUtil.randomBytes(4096));
+    Buffer buffer = new Buffer(io.vertx.core.buffer.Buffer.buffer(ByteUtil.randomBytes(4096)));
 
     // write data to the stream
     inputStream.write(buffer);
     Assert.assertFalse(inputStream.writeQueueFull());
 
     // write more data to the stream: 8191 < 8192
-    buffer = Buffer.buffer(ByteUtil.randomBytes(4095));
+    buffer = new Buffer(io.vertx.core.buffer.Buffer.buffer(ByteUtil.randomBytes(4095)));
     inputStream.write(buffer);
     Assert.assertFalse(inputStream.writeQueueFull());
 
     // queue should be full (8192 bytes in queue)
-    buffer = Buffer.buffer(ByteUtil.randomBytes(1));
+    buffer = new Buffer(io.vertx.core.buffer.Buffer.buffer(ByteUtil.randomBytes(1)));
     inputStream.write(buffer);
     Assert.assertTrue(inputStream.writeQueueFull());
   }
@@ -132,7 +131,7 @@ public class GridFSInputStreamTest {
   public void testDrainHandler() {
     GridFSInputStream inputStream = GridFSInputStream.create();
 
-    Buffer buffer = Buffer.buffer(ByteUtil.randomBytes(8192));
+    Buffer buffer = new Buffer(io.vertx.core.buffer.Buffer.buffer(ByteUtil.randomBytes(8192)));
 
     // write more data than the queue has capacity
     inputStream.write(buffer);
@@ -192,7 +191,7 @@ public class GridFSInputStreamTest {
 
     inputStream.setWriteQueueMaxSize(2);
 
-    Buffer buffer = Buffer.buffer(ByteUtil.randomBytes(1));
+    Buffer buffer = new Buffer(io.vertx.core.buffer.Buffer.buffer(ByteUtil.randomBytes(1)));
     inputStream.write(buffer);
 
     Assert.assertFalse(inputStream.writeQueueFull());
